requirements.txt
python-dotenv
pyqt6
PyQt6-WebEngine
geopy
folium